+++ engine/runtime-parser/src/main/scala/org/enso/compiler/core/ir/expression/errors/ImportExport.scala
@@ -171,6 +171,38 @@ object ImportExport {
       s"No such constructor ${constructorName} in type $typeName"
   }
 
+  case class ExportSymbolsFromPrivateModule(
+    moduleName: String
+  ) extends Reason {
+    override def message: String =
+      s"Cannot export any symbol from module '$moduleName': The module is private"
+  }
+
+  case class ExportPrivateModule(
+    moduleName: String
+  ) extends Reason {
+    override def message: String =
+      s"Cannot export private module '$moduleName'"
+  }
+
+  case class ImportPrivateModule(
+    moduleName: String
+  ) extends Reason {
+    override def message: String =
+      s"Cannot import private module '$moduleName'"
+  }
+
+  case class SubmoduleVisibilityMismatch(
+    moduleName: String,
+    submoduleName: String,
+    moduleVisibility: String,
+    submoduleVisibility: String
+  ) extends Reason {
+    override def message: String =
+      s"Cannot export submodule '$submoduleName' of module '$moduleName': " +
+      s"the submodule is $submoduleVisibility, but the module is $moduleVisibility"
+  }
+
   /** Represents an ambiguous import resolution error, where the same symbol is imported more than once refereing
     * to different objects. The objects are represented by their physical path in the project.
     *