+++ engine/runtime/src/main/java/org/enso/interpreter/util/ScalaConversions.java
@@ -33,4 +33,9 @@ public class ScalaConversions {
   public static <T> scala.collection.immutable.List<T> nil() {
     return (scala.collection.immutable.List<T>) scala.collection.immutable.Nil$.MODULE$;
   }
+
+  public static <T> scala.collection.immutable.List<T> cons(
+      T head, scala.collection.immutable.List<T> tail) {
+    return scala.collection.immutable.$colon$colon$.MODULE$.apply(head, tail);
+  }
 }