+++ engine/runtime/src/test/scala/org/enso/interpreter/test/semantic/ImportsTest.scala
@@ -219,4 +219,41 @@ class ImportsTest extends PackageTest {
     outLines(2) shouldEqual "(D_Mod.Value 1)"
   }
 
+  "Private modules" should "be able to import and use private modules within the same project" in {
+    evalTestProject(
+      "Test_Private_Modules_1"
+    ).toString shouldEqual "42"
+  }
+
+  "Private modules" should "be able to import non-private stuff" in {
+    evalTestProject(
+      "Test_Private_Modules_2"
+    ).toString shouldEqual "(Pub_Mod_Type.Value 42)"
+  }
+
+  "Private modules" should "not be able to import private modules from different project" in {
+    the[InterpreterException] thrownBy evalTestProject(
+      "Test_Private_Modules_3"
+    ) should have message "Compilation aborted due to errors."
+    val outLines = consumeOut.filterNot(isDiagnosticLine)
+    outLines should have length 1
+    outLines.head should include(
+      "Main.enso:2:1: error: Cannot import private module"
+    )
+  }
+
+  "Private modules" should "not be able to mix private and public submodules" in {
+    val e = the[InterpreterException] thrownBy evalTestProject(
+      "Test_Private_Modules_4"
+    )
+    e.getMessage() should include(
+      "Cannot export submodule 'local.Test_Private_Modules_4.Sub.Priv_SubMod' of module 'local.Test_Private_Modules_4.Sub'"
+    )
+  }
+
+  "Private module" should "be able to have only private submodules" in {
+    evalTestProject(
+      "Test_Private_Modules_5"
+    ) shouldEqual 42
+  }
 }