+++ CHANGELOG.md
@@ -972,6 +972,7 @@
 - [Downloadable VSCode extension][7861]
 - [New `project/status` route for reporting LS state][7801]
 - [Add Enso-specific assertions][7883])
+- [Modules can be `private`][7840]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -1117,6 +1118,7 @@
 [7825]: https://github.com/enso-org/enso/pull/7825
 [7861]: https://github.com/enso-org/enso/pull/7861
 [7883]: https://github.com/enso-org/enso/pull/7883
+[7840]: https://github.com/enso-org/enso/pull/7840
 
 # Enso 2.0.0-alpha.18 (2021-10-12)